pull out node tests, add tests for privateEncrypt publicDecrypt
