+++ index.js
@@ -7,7 +7,12 @@ exports.privateDecrypt = crypto.privateDecrypt;
 
 if (typeof crypto.privateEncrypt !== 'function') {
   exports.privateEncrypt = require('./browser').privateEncrypt;
+} else {
+  exports.privateEncrypt = crypto.privateEncrypt;
 }
+
 if (typeof crypto.publicDecrypt !== 'function') {
   exports.publicDecrypt = require('./browser').publicDecrypt;
+} else {
+  exports.publicDecrypt = crypto.publicDecrypt;
 }
